refactor: Add common folder
