+++ lib/home/widgets/chart.dart
@@ -1,7 +1,7 @@
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 
-import '../../constants.dart';
+import '../../common/constants.dart';
 
 class Chart extends StatelessWidget {
   const Chart({Key? key}) : super(key: key);
