+++ lib/home/widgets/resource_info_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-import '../../constants.dart';
+import '../../common/constants.dart';
 
 class ResourceInfoCard extends StatelessWidget {
   const ResourceInfoCard({
