+++ lib/home/widgets/resource_usage_details.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../constants.dart';
+import '../../common/constants.dart';
 import 'chart.dart';
 import 'resource_info_card.dart';
